feat: extract Firebase operations from PropertyDetails component
- Create propertyDetailsService.js with all Firebase operations
- Extract loadPropertyDetails, getSampleImages, getLocationText, getPriceText, getPropertyTypeDisplay, and generateShareMessage functions
- Update PropertyDetails.jsx to use service functions instead of direct Firebase calls
- Improve separation of concerns and maintainability
- Follow existing service architecture pattern
